refactor(server): split config-center.ts into domain/config-center/ (part of #1563)
The 4622-line apps/server/src/config-center.ts is split into 12 sibling
files under apps/server/src/domain/config-center/, each under 1000 lines.
The original config-center.ts becomes a one-line re-export shim so all
33 existing call sites continue working without modification.

New layout:
- types.ts         (384) types, interfaces, public API shapes
- constants.ts     (234) CONFIG_DEFINITIONS, impact rules, preset ids
- schemas.ts       (609) CONFIG_DOCUMENT_SCHEMAS + schema helpers
- helpers.ts       (335) small utilities (pushIssue, parseJsonPath, etc.)
- diff.ts          (200) diff/impact/preview builders
- presets.ts       (429) builtin preset applicators
- workbook.ts      (113) xlsx/csv/jsonc import/export
- preview.ts       (270) createWorldConfigPreview + parseConfigDocument
- runtime.ts       (328) runtime bundle apply/rollback/hot-reload
- validators.ts    (562) detailed per-document validators
- store.ts         (955) BaseConfigCenterStore + File/MySql subclasses
- routes.ts        (520) registerConfigCenterRoutes
- index.ts         (34)  barrel re-exports

One behavioral touch: moved the inline init-runtime state assignment
block in BaseConfigCenterStore.initializeRuntimeConfigs into
runtime.ts#initializeAppliedRuntimeBundle so module-level state stays
encapsulated in runtime.ts.

Refs #1563
